requirements.txt
langgraph
langchain-google-genai
langchain-community
langchain-tavily
motor
pymongo
tavily-python
docling
youtube-transcript-api
streamlit
python-dotenv
langgraph-checkpoint-mongodb
langchain-huggingface
sentence-transformers
faiss-cpu
yt-dlp